--story=117115687 【云桌面】封装一个openapiapi，创建离岸团队项目的云桌面接口，给devcloud调用
--story=117182231 【云桌面】devcloud接口优化需求
--story=117182032 【云桌面】增加待分配的实例通知
--story=117138428 【云桌面】创建个人云桌面时创建个人preci项目，同时调用bkrepo的接口做初始化

+++ src/backend/ci/ext/tencent/remotedev/biz-remotedev-tencent/src/main/kotlin/com/tencent/devops/remotedev/resources/service/ServiceRemoteDevResourceImpl.kt
@@ -16,6 +16,8 @@ import com.tencent.devops.remotedev.common.Constansts
 import com.tencent.devops.remotedev.pojo.WindowsResourceTypeConfig
 import com.tencent.devops.remotedev.pojo.WindowsWorkspaceCreate
 import com.tencent.devops.remotedev.pojo.WorkspaceOwnerType
+import com.tencent.devops.remotedev.pojo.async.AsyncPipelineEvent
+import com.tencent.devops.remotedev.pojo.common.QuotaType
 import com.tencent.devops.remotedev.pojo.op.OpProjectWorkspaceAssignData
 import com.tencent.devops.remotedev.pojo.op.RemotedevCvmData
 import com.tencent.devops.remotedev.pojo.op.WorkspaceNotifyData
@@ -239,10 +241,15 @@ class ServiceRemoteDevResourceImpl(
     }
 
     override fun createPersonalWorkspace(userId: String, data: WindowsWorkspaceCreate): Result<Boolean> {
-        return Result(createControl.devcloudCreateWorkspace(userId, data))
+        return Result(createControl.devcloudCreateWorkspace(userId = userId, workspaceCreate = data, projectId = null))
     }
 
     override fun deletePersonalWorkspace(userId: String, workspaceName: String): Result<Boolean> {
+        val record = workspaceService.getWorkspaceRecord(workspaceName = workspaceName)
+        if (record == null || record.ownerType != WorkspaceOwnerType.PERSONAL) {
+            logger.warn("delete personal workspace with invalid workspace type: $userId|$workspaceName")
+            return Result(false)
+        }
         return Result(
             deleteControl.deleteWorkspace(
                 userId = userId,
@@ -254,11 +261,67 @@ class ServiceRemoteDevResourceImpl(
     }
 
     override fun getPersonalWorkspace(userId: String, workspaceName: String): Result<WeSecProjectWorkspace?> {
+        val record = workspaceService.getWorkspaceRecord(workspaceName = workspaceName)
+        if (record == null || record.ownerType != WorkspaceOwnerType.PERSONAL) {
+            logger.warn("get personal workspace with invalid workspace type: $userId|$workspaceName")
+            return Result(null)
+        }
         return Result(
             workspaceService.getWorkspaceList4WeSec(
                 workspaceName = workspaceName,
                 notStatus = null
             ).firstOrNull()
         )
+    }
+
+    override fun createProjectWorkspace(
+        userId: String,
+        projectId: String,
+        data: WindowsWorkspaceCreate
+    ): Result<Boolean> {
+        permissionService.checkUserManager(userId, projectId)
+        return Result(
+            createControl.devcloudCreateWorkspace(userId = userId, workspaceCreate = data, projectId = projectId)
+        )
+    }
+
+    override fun deleteProjectWorkspace(userId: String, projectId: String, workspaceName: String): Result<Boolean> {
+        val record = workspaceService.getWorkspaceRecord(workspaceName = workspaceName)
+        if (record == null || record.ownerType != WorkspaceOwnerType.PROJECT || record.projectId != projectId) {
+            logger.warn("delete project workspace with invalid workspace type: $userId|$projectId|$workspaceName")
+            return Result(false)
+        }
+
+        return Result(
+            deleteControl.deleteWorkspace(
+                userId = userId,
+                workspaceName = workspaceName,
+                needPermission = !permissionService.hasUserManager(userId, projectId),
+                checkDeleteImmediately = true
+            )
+        )
+    }
+
+    override fun getProjectWorkspace(
+        userId: String,
+        projectId: String,
+        workspaceName: String
+    ): Result<WeSecProjectWorkspace?> {
+        val record = workspaceService.getWorkspaceRecord(workspaceName = workspaceName)
+        if (record == null || record.ownerType != WorkspaceOwnerType.PROJECT || record.projectId != projectId) {
+            logger.warn("get project workspace with invalid workspace type: $userId|$projectId|$workspaceName")
+            return Result(null)
+        }
+        permissionService.checkViewerPermission(userId, workspaceName, projectId)
+        return Result(
+            workspaceService.getWorkspaceList4WeSec(
+                workspaceName = workspaceName,
+                notStatus = null
+            ).firstOrNull()
+        )
+    }
+
+    override fun getWindowsQuota(userId: String, type: QuotaType): Result<Map<String, Map<String, Int>>> {
+        return Result(windowsResourceConfigService.allWindowsQuota(userId, false, type))
     }
 }
